Save and load implemented for Tuesday, Wednesday, Thursday, and Friday
Save and loading feature now complete.

Possible features to implement:
Drop down boxes for setting time and date for notes, activities, and
classes

Make a JOptionPane pop up if the user has not  completely filled up the
parts of the note, activity, or class

Make a JOptionPane pop up when saving.